Bug in block provider
When the orderer provides a block that does not pass validation,
the client did not disconnect, which leads to an livelock,
since all future (higher) blocks from the same orderer will be
stalled due to the gap, and the client will not ask for the same
block again. The solution: to disconnect from the offending
orderer and  try again (seek) from last good block.

Change-Id: I230203b9f450f47362db608f3c392236e2090fbf
Signed-off-by: Yoav Tock <tock@il.ibm.com>